schemas in child classes
