--- src/opengeodeweb_viewer/object/schemas/set_color.json
@@ -1,39 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "id": {
-      "type": "string"
-    },
-    "color": {
-      "type": "object",
-      "properties": {
-        "red": {
-          "type": "integer",
-          "minimum": 0,
-          "maximum": 255
-        },
-        "green": {
-          "type": "integer",
-          "minimum": 0,
-          "maximum": 255
-        },
-        "blue": {
-          "type": "integer",
-          "minimum": 0,
-          "maximum": 255
-        }
-      },
-      "required": [
-        "red",
-        "green",
-        "blue"
-      ],
-      "additionalProperties": false
-    }
-  },
-  "required": [
-    "id",
-    "color"
-  ],
-  "additionalProperties": false
-}
