+++ src/opengeodeweb_viewer/rpc/mesh/schemas/apply_textures.json
