+++ src/opengeodeweb_viewer/rpc/mesh/schemas/deregister.json
@@ -1,16 +1,13 @@
 {
+  "rpc": "mesh.deregister",
   "type": "object",
   "properties": {
     "id": {
       "type": "string"
-    },
-    "visibility": {
-      "type": "boolean"
     }
   },
   "required": [
-    "id",
-    "visibility"
+    "id"
   ],
   "additionalProperties": false
 }
