+++ src/opengeodeweb_viewer/rpc/model/schemas/register.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "object.register",
+  "rpc": "model.register",
   "type": "object",
   "properties": {
     "id": {
