--- src/opengeodeweb_viewer/rpc/model/schemas/set_model_mesh_visibility.json
@@ -1,17 +1,0 @@
-{
-  "rpc": "set_model_mesh_visibility",
-  "type": "object",
-  "properties": {
-    "id": {
-      "type": "string"
-    },
-    "visibility": {
-      "type": "boolean"
-    }
-  },
-  "required": [
-    "id",
-    "visibility"
-  ],
-  "additionalProperties": false
-}
