+++ src/opengeodeweb_viewer/rpc/viewer/schemas/create_visualization.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "create_visualization",
+  "rpc": "viewer.create_visualization",
   "type": "object",
   "properties": {},
   "required": [],
