+++ src/opengeodeweb_viewer/rpc/viewer/schemas/set_background_color.json
@@ -1,5 +1,5 @@
 {
-    "rpc": "set_viewer_background_color",
+    "rpc": "viewer.set_background_color",
     "type": "object",
     "properties": {
         "red": {
