+++ src/opengeodeweb_viewer/rpc/viewer/schemas/update_data.json
@@ -1,11 +1,13 @@
 {
-  "rpc": "update_data",
+  "rpc": "viewer.update_data",
   "type": "object",
   "properties": {
     "id": {
       "type": "string"
     }
   },
-  "required": ["id"],
+  "required": [
+    "id"
+  ],
   "additionalProperties": false
 }
