+++ src/tests/data/images/mesh/apply_textures.jpeg
