+++ src/tests/data/images/mesh/deregister.jpeg
