+++ src/tests/data/images/mesh/register.jpeg
