+++ src/tests/data/images/mesh/set_color.jpeg
