+++ src/tests/data/images/mesh/set_opacity.jpeg
