+++ src/tests/data/images/mesh/set_point_size_1.jpeg
