+++ src/tests/data/images/mesh/toggle_object_visibility_2.jpeg
