+++ src/tests/data/images/mesh/toggle_object_visiblity.jpeg
