+++ src/tests/data/images/mesh/toggle_point_visibility.jpeg
