+++ src/tests/data/images/viewer/set_background_color.jpeg
